tests propogate failure, more expressive spec, support vscode fixtures
