+++ test/specs/test.cpp.json
@@ -51,7 +51,7 @@
         "scopes": ["punctuation.definition.comment"]
     },
     {
-        "source": "a comment /* no nest */",
+        "source": " a comment /* no nest */",
         "scopesEnd": ["comment.line.double-slash"]
     }
 ]
